+++ targets/TARGET_NXP/TARGET_MCUXpresso_MCUS/TARGET_MIMXRT1050/TARGET_EVK/TARGET_1050_EVK/xip/evkbimxrt1050_sdram_ini_dcd.c
@@ -238,9 +238,9 @@ const uint8_t dcd_data[] = {
     0x40, 0x1F, 0x82, 0xA8, 0x00, 0x01, 0x10, 0xF9,
     /* #1.95, command: write_value, address: SEMC_MCR, value: 0x10000004, size: 4 */
     0x40, 0x2F, 0x00, 0x00, 0x10, 0x00, 0x00, 0x04,
-    /* #1.96, command: write_value, address: SEMC_BMCR0, value: 0x30524, size: 4 */
+    /* #1.96, command: write_value, address: SEMC_BMCR0, value: 0x81, size: 4 */
     0x40, 0x2F, 0x00, 0x08, 0x00, 0x00, 0x00, 0x81,
-    /* #1.97, command: write_value, address: SEMC_BMCR1, value: 0x6030524, size: 4 */
+    /* #1.97, command: write_value, address: SEMC_BMCR1, value: 0x81, size: 4 */
     0x40, 0x2F, 0x00, 0x0C, 0x00, 0x00, 0x00, 0x81,
     /* #1.98, command: write_value, address: SEMC_BR0, value: 0x8000001B, size: 4 */
     0x40, 0x2F, 0x00, 0x10, 0x80, 0x00, 0x00, 0x1B,
