+++ targets/targets.json
@@ -5085,7 +5085,8 @@
             "STDIO_MESSAGES",
             "TRNG",
             "WATCHDOG",
-            "USBDEVICE"
+            "USBDEVICE",
+            "LPTICKER"
         ],
         "features": [
             "LWIP"
@@ -5093,8 +5094,14 @@
         "device_name": "MIMXRT1052",
         "overrides": {
             "deep-sleep-latency": 10
-        }
-    },
+        },
+         "config": {
+             "enable-overdrive-mode": {
+                 "help": "Enable 'overdrive mode' and raise core clock from 528MHz to 600MHz.  Only allowed on MIMXRT10xxD parts, not MIMXRT10xxC",
+                 "value": 1
+             }
+         }
+     },
     "MIMXRT1050_EVK": {
         "inherits": [
             "MIMXRT105X"
@@ -5110,10 +5117,9 @@
             "XIP_BOOT_HEADER_DCD_ENABLE=1",
             "FSL_FEATURE_PHYKSZ8081_USE_RMII50M_MODE",
             "MIMXRT105X_BOARD_HAS_EXTERNAL_RAM=1",
-            "MIMXRT105X_EXTERNAL_RAM_SIZE=0x01E00000"
+            "MIMXRT105X_EXTERNAL_RAM_SIZE=0x02000000"
         ],
         "device_has_add": [
-            "LPTICKER",
             "FLASH"
         ],
         "detect_code": [
